+++ databasez/core/database.py
@@ -11,7 +11,7 @@ from functools import lru_cache
 from types import TracebackType
 
 from databasez import interfaces
-from databasez.utils import thread_protector
+from databasez.utils import multiloop_protector
 
 from .connection import Connection
 from .databaseurl import DatabaseURL
@@ -242,7 +242,7 @@ class Database:
     async def connect_hook(self) -> None:
         """Refcount protected connect hook. Executed begore engine and global connection setup."""
 
-    @thread_protector(True)
+    @multiloop_protector(True)
     async def connect(self) -> bool:
         """
         Establish the connection pool.
@@ -269,7 +269,7 @@ class Database:
     async def disconnect_hook(self) -> None:
         """Refcount protected disconnect hook. Executed after connection, engine cleanup."""
 
-    @thread_protector(True)
+    @multiloop_protector(True)
     async def disconnect(self, force: bool = False) -> bool:
         """
         Close all connections in the connection pool.
@@ -330,7 +330,7 @@ class Database:
                 database = dbs.pop(loop, database)
         await database.disconnect()
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def fetch_all(
         self,
         query: typing.Union[ClauseElement, str],
@@ -339,7 +339,7 @@ class Database:
         async with self.connection() as connection:
             return await connection.fetch_all(query, values)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def fetch_one(
         self,
         query: typing.Union[ClauseElement, str],
@@ -350,7 +350,7 @@ class Database:
             return await connection.fetch_one(query, values, pos=pos)
             assert connection._connection_counter == 1
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def fetch_val(
         self,
         query: typing.Union[ClauseElement, str],
@@ -361,7 +361,7 @@ class Database:
         async with self.connection() as connection:
             return await connection.fetch_val(query, values, column=column, pos=pos)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def execute(
         self,
         query: typing.Union[ClauseElement, str],
@@ -370,14 +370,14 @@ class Database:
         async with self.connection() as connection:
             return await connection.execute(query, values)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def execute_many(
         self, query: typing.Union[ClauseElement, str], values: typing.Any = None
     ) -> typing.Union[typing.Sequence[interfaces.Record], int]:
         async with self.connection() as connection:
             return await connection.execute_many(query, values)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def iterate(
         self,
         query: typing.Union[ClauseElement, str],
@@ -388,7 +388,7 @@ class Database:
             async for record in connection.iterate(query, values, chunk_size):
                 yield record
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def batched_iterate(
         self,
         query: typing.Union[ClauseElement, str],
@@ -400,11 +400,11 @@ class Database:
             async for records in connection.batched_iterate(query, values, batch_size):
                 yield batch_wrapper(records)
 
-    @thread_protector(True)
+    @multiloop_protector(True)
     def transaction(self, *, force_rollback: bool = False, **kwargs: typing.Any) -> "Transaction":
         return Transaction(self.connection, force_rollback=force_rollback, **kwargs)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def run_sync(
         self,
         fn: typing.Callable[..., typing.Any],
@@ -414,17 +414,17 @@ class Database:
         async with self.connection() as connection:
             return await connection.run_sync(fn, *args, **kwargs)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def create_all(self, meta: MetaData, **kwargs: typing.Any) -> None:
         async with self.connection() as connection:
             await connection.create_all(meta, **kwargs)
 
-    @thread_protector(False)
+    @multiloop_protector(False)
     async def drop_all(self, meta: MetaData, **kwargs: typing.Any) -> None:
         async with self.connection() as connection:
             await connection.drop_all(meta, **kwargs)
 
-    @thread_protector(False, wrap_context_manager=True)
+    @multiloop_protector(False, wrap_context_manager=True)
     def connection(self) -> Connection:
         if self.force_rollback:
             return typing.cast(Connection, self._global_connection)
@@ -435,7 +435,7 @@ class Database:
         return self._connection
 
     @property
-    @thread_protector(True)
+    @multiloop_protector(True)
     def engine(self) -> typing.Optional[AsyncEngine]:
         return self.backend.engine